+++ src/public/js/views/competitorPromotion/preView/preView.js
@@ -23,11 +23,12 @@ define([
     'constants/errorMessages',
     'constants/aclRoleIndexes',
     'views/competitorPromotion/edit',
-    'models/competitorPromotion'
+    'models/competitorPromotion',
+    'constants/infoMessages'
 ], function (Backbone, _, $, moment, PreviewTemplate, FileTemplate, CommentTemplate, NewCommentTemplate,
              FileCollection, FileModel, CommentModel, BaseView, CommentCollection,
              populate, CONSTANTS, levelConfig, implementShowHideArabicInputIn, dataService, CONTENT_TYPES,
-             FileDialogView, FileDialogPreviewView, ERROR_MESSAGES, ACL_ROLES, EditView, CompetitorPromotionModel) {
+             FileDialogView, FileDialogPreviewView, ERROR_MESSAGES, ACL_ROLES, EditView, CompetitorPromotionModel, INFO_MESSAGES) {
 
     var PreView = BaseView.extend({
         contentType: CONTENT_TYPES.COMPETITORPROMOTION,
@@ -88,12 +89,28 @@ define([
                 model.on('competitor-promotion-edited', function (response) {
                     var view = that.$el;
 
+                    response.dateStart = moment.utc(response.dateStart).format('DD.MM.YYYY');
+                    response.dateEnd = moment.utc(response.dateEnd).format('DD.MM.YYYY');
+
+                    that.model.set(response, {
+                        merge: true
+                    });
+
                     view.find('#promotion').html(data.promotion);
                     view.find('#price').html(data.price);
                     view.find('#packing').html(data.packing);
-                    view.find('#expiry').html(moment(data.expiry).format('DD.MM.YYYY'));
-                    view.find('#date-start').html(moment(data.dateStart).format('DD.MM.YYYY'));
-                    view.find('#date-end').html(moment(data.dateEnd).format('DD.MM.YYYY'));
+
+                    if (data.expiry) {
+                        view.find('#expiry').html(moment.utc(data.expiry).format('DD.MM.YYYY'));
+                    }
+
+                    if (data.dateStart) {
+                        view.find('#date-start').html(moment.utc(data.dateStart).format('DD.MM.YYYY'));
+                    }
+
+                    if (data.dateEnd) {
+                        view.find('#date-end').html(moment.utc(data.dateEnd).format('DD.MM.YYYY'));
+                    }
 
                     var displayTypeString = response.displayType.map(function (item) {
                         return item.name[App.currentUser.currentLanguage];
@@ -109,16 +126,18 @@ define([
         },
 
         deleteCompetitorPromotion: function () {
-            var  that = this;
-            var model = new CompetitorPromotionModel();
+            if (confirm(INFO_MESSAGES.confirmDeleteCompetitorPromotionActivity[App.currentUser.currentLanguage])) {
+                var  that = this;
+                var model = new CompetitorPromotionModel();
 
-            model.delete(that.model.get('_id'));
+                model.delete(that.model.get('_id'));
 
-            model.on('competitor-promotion-deleted', function () {
-                that.trigger('update-list-view');
+                model.on('competitor-promotion-deleted', function () {
+                    that.trigger('update-list-view');
 
-                that.$el.dialog('close').dialog('destroy').remove();
-            });
+                    that.$el.dialog('close').dialog('destroy').remove();
+                });
+            }
         },
 
         showFilePreviewDialog: _.debounce(function (e) {
@@ -126,18 +145,25 @@ define([
             var $thumbnail = $el.closest('.masonryThumbnail');
             var fileModelId = $thumbnail.attr('data-id');
             var fileModel = this.previewFiles.get(fileModelId);
-
-            this.fileDialogView = new FileDialogPreviewView({
-                fileModel  : fileModel,
-                bucket     : this.contentType,
+            var options = {
+                fileModel: fileModel,
                 translation: this.translation
-            });
+            };
+
+            if (!fileModel) {
+                fileModel = this.fileCollection.get(fileModelId);
+                options.fileModel = fileModel;
+            }
+
+            this.fileDialogView = new FileDialogPreviewView(options);
             this.fileDialogView.on('download', function (options) {
                 var url = options.url;
                 var originalName = options.originalName;
-                var $fileElement;
+
                 $thumbnail.append('<a class="hidden" id="downloadFile" href="' + url + '" download="' + originalName + '"></a>');
-                $fileElement = $thumbnail.find('#downloadFile');
+
+                var $fileElement = $thumbnail.find('#downloadFile');
+
                 $fileElement[0].click();
                 $fileElement.remove();
             });
@@ -259,7 +285,7 @@ define([
                         }
 
                         $target.addClass('loadFile');
-
+                        self.fileCollection = new FileCollection(filesCollection);
                         self.showFilesInComment($showFilesBlock, filesCollection);
                     });
                 } else {
